+++ Modules/Intent.Modules.AspNetCore/Templates/ProblemDetailsConfiguration/ProblemDetailsConfigurationTemplatePartial.cs
@@ -10,7 +10,6 @@ using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.VisualStudio;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
-using static System.Runtime.InteropServices.JavaScript.JSType;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
@@ -104,7 +103,6 @@ namespace Intent.Modules.AspNetCore.Templates.ProblemDetailsConfiguration
                 .AddUsing("Microsoft.AspNetCore.Hosting")
                 .AddUsing("Microsoft.AspNetCore.Diagnostics")
                 .AddUsing("System.Collections.Generic");
-
             method.AddInvocationStatement("services.AddProblemDetails", addProblemDetails => addProblemDetails
                     .AddArgument(new CSharpLambdaBlock("conf")
                         .WithExpressionBody(new CSharpLambdaBlock("conf.CustomizeProblemDetails = context")