+++ Modules/Intent.Modules.AspNetCore.Controllers.Dispatch.ServiceContract/release-notes.md
@@ -1,6 +1,6 @@
 ### Version 5.2.9
 
-- Improvement: Small updated to more align code styling with best practices
+- Improvement: Small updated to module internal code
 
 ### Version 5.2.8