+++ Modules/Intent.Modules.Eventing.MassTransit/release-notes.md
@@ -1,6 +1,6 @@
 ### Version 6.2.7
 
-- Improvement: Small updated to more align code styling with best practices
+- Improvement: Small updated to module internal code
 
 ### Version 6.2.6